+++ src/components/TextField/stories/SearchField.stories.tsx
@@ -1,12 +1,20 @@
 import React from "react";
 import { Meta, StoryFn } from "@storybook/react";
 
+import { helpersArgTypes } from "../../../../.storybook/helpers";
+
 import SearchField from "./../SearchField";
 
 export default {
     title: "Components/SearchField",
     component: SearchField,
     argTypes: {
+        leftIcon: {
+            ...helpersArgTypes.exampleIcon,
+        },
+        rightElement: {
+            ...helpersArgTypes.exampleIcon,
+        },
         hasStatePrimary: { table: { disable: true } },
         hasStateSuccess: { table: { disable: true } },
         hasStateWarning: { table: { disable: true } },
@@ -23,7 +31,7 @@ Default.args = {
     onClearanceText: "",
 };
 
-export const SearchFieldWithClearanceIcon: StoryFn<typeof SearchField> = (args) => {
+const SearchFieldWithClearanceIconTemplate: StoryFn<typeof SearchField> = (args) => {
     const [query, setQuery] = React.useState<string>("");
     return (
         <SearchField
@@ -33,4 +41,10 @@ export const SearchFieldWithClearanceIcon: StoryFn<typeof SearchField> = (args) 
             onClearanceHandler={() => setQuery("")}
         />
     );
+};
+
+export const SearchFieldWithClearanceIcon = SearchFieldWithClearanceIconTemplate.bind({});
+SearchFieldWithClearanceIcon.args = {
+    onClearanceHandler: null,
+    onClearanceText: "",
 };